# backend/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: HireSense (知聘) - Backend
  description: HireSense 后端接口，涵盖了用户鉴权、档案管理、实时沉浸式面试交互以及多维度面试报告生成等核心能力。
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发环境的服务地址
security:
  - bearerAuth: []
...
paths:
...
  /api/user/password:
    put:
      tags: [User]
      summary: 修改密码
      description: 更新当前登录用户的账户密码。
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserPasswordRequest" }
      responses:
        "200":
          description: 密码修改成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPasswordResponse" }
  /api/user/profile:
    get:
      tags: [User]
      summary: 获取用户档案
      description: 读取当前求职者的基本信息、简历内容以及 AI 面试官倾向设置。
      responses:
        "200":
          description: 获取档案成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileResponse" }
    put:
      tags: [User]
      summary: 更新用户档案
      description: 覆写求职者档案，例如更新期望岗位、打磨后的简历以及面试难度配置。
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Profile" }
      responses:
        "200":
          description: 档案更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PutProfileResponse" }
...
  /api/user/favorite-questions:
    get:
      tags: [User]
      summary: 获取收藏题列表
      description: 读取当前登录用户收藏的面试官问题，按最近一次收藏时间倒序返回。
      responses:
        "200":
          description: 成功返回收藏题列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteQuestionsResponse" }
    post:
      tags: [User]
      summary: 收藏一道题
      description: 基于面试报告中的某条复盘记录，将对应面试官问题加入收藏题列表。
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddFavoriteQuestionRequest" }
      responses:
        "200":
          description: 收藏成功并返回所属收藏题分组
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteQuestionResponse" }
  /api/user/favorite-questions/source/{interviewID}/{reviewIndex}:
    delete:
      tags: [User]
      summary: 移除收藏来源
      description: 从收藏题中移除指定报告来源；若该题已无任何来源，则整题收藏会被自动删除。
      parameters:
        - name: interviewID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: reviewIndex
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: 收藏来源移除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteFavoriteQuestionSourceResponse" }
  /api/user/favorite-questions/{favoriteID}:
    delete:
      tags: [User]
      summary: 删除整题收藏
      description: 删除某道收藏题及其所有关联来源。
      parameters:
        - name: favoriteID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 收藏题移除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteFavoriteQuestionResponse" }

  # ==========================================
  # Interview
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "传递形式: `Authorization: Bearer <Your_Token>`"
  schemas:
...
    UserPasswordRequest:
      type: object
      properties:
        password:
          type: string
          description: 新密码
          examples:
            - "N3wStrong!P@ss"
      required: [password]
    UserPasswordResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "密码修改成功"
        data: {}
      required: [success, message, data]
    Profile:
      type: object
      description: 用户详细职业档案和 AI 调优设定
      properties:
        name:
          type: string
          description: 姓名
          examples:
            - "李工"
        job:
          type: string
          enum: ["frontend", "backend"]
          description: 投递或期望的职位角色
        resume:
          type: string
          description: 结构化的简历文本
          examples:
            - "5 年开发经验。精通 React/Vue 生态体系，曾主导并重构千万级 PV 的大型后台系统。熟悉 Webpack 等工程化构建体系。"
        personalization:
          type: string
          description: 专门下达给 AI 面试官的人设或追问偏好指令
          examples:
            - "请扮演字节跳动的高级技术总监，侧重在回答中持续深挖高并发、性能优化等知识点，并且态度保持客观严厉。"
      required: [name, job]
    ProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "读取档案成功"
        data:
          $ref: "#/components/schemas/Profile"
      required: [success, message, data]
    PutProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "档案覆写成功"
        data: {}
      required: [success, message, data]
...
    AddFavoriteQuestionRequest:
      type: object
      properties:
        interview_id:
          type: string
          format: uuid
          description: 收藏来源所属的面试 ID
        review_index:
          type: integer
          minimum: 0
          description: 对应报告 reviews 数组中的索引
      required: [interview_id, review_index]
    FavoriteQuestionsResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "读取收藏题成功"
        data:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteQuestion"
      required: [success, message, data]
    FavoriteQuestionResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "收藏题保存成功"
        data:
          $ref: "#/components/schemas/FavoriteQuestion"
      required: [success, message, data]
    DeleteFavoriteQuestionSourceResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "收藏来源移除成功"
        data: {}
      required: [success, message, data]
    DeleteFavoriteQuestionResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "收藏题移除成功"
        data: {}
      required: [success, message, data]
    FavoriteQuestion:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 收藏题分组唯一 ID
        question:
          type: string
          description: 去重后的题目文本
        created_at:
          type: string
          format: date-time
          description: 首次收藏该题的时间
        updated_at:
          type: string
          format: date-time
          description: 最近一次新增来源的时间
        source_count:
          type: integer
          minimum: 1
          description: 该题当前保留的来源数量
        sources:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteQuestionSource"
      required: [id, question, created_at, updated_at, source_count, sources]
    FavoriteQuestionSource:
      type: object
      properties:
        interview_id:
          type: string
          format: uuid
          description: 该收藏来源所属的面试 ID
        review_index:
          type: integer
          minimum: 0
          description: 对应报告 reviews 中的索引
        interview_created_at:
          type: number
          description: 来源面试的创建时间戳（毫秒）
        favorited_at:
          type: string
          format: date-time
          description: 该来源被加入收藏的时间
        interviewer:
          type: string
          description: 面试官问题
        interviewer_role:
          $ref: "#/components/schemas/InterviewSpeakerRole"
        interviewee:
          type: string
          description: 求职者当时的实际回答
        standard_answer:
          type: string
          description: 当前题目的标准答案
        score:
          type: number
          description: 该轮得分
        advice:
          type: string
          description: AI 点评与纠误建议
        concept:
          type: string
          description: 当前题目的考点
        difficulty_label:
          type: string
          description: 当前题目的难度标签
        score_breakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        reason_tags:
          type: array
          items:
            type: string
          description: 支撑该分数的机器标签
        strength_points:
          type: array
          items:
            type: string
          description: 本题回答中已经答到的点
        missing_points:
          type: array
          items:
            type: string
          description: 本题回答中缺失的点
        score_rationale:
          type: string
          description: 对题级分数的结构化解释
      required:
        [
          interview_id,
          review_index,
          interview_created_at,
          favorited_at,
          interviewer,
          interviewer_role,
          interviewee,
          score,
          advice,
          difficulty_label,
          reason_tags,
          strength_points,
          missing_points,
        ]
...
    InterviewSpeakerRole:
      type: string
      description: 发言的 AI 面试官角色
      enum: ["ai", "hr", "tech_lead", "executive"]
      examples:
        - "tech_lead"
...
    ScoreBreakdown:
      type: object
      description: 题级分项得分
      properties:
        coverage_score:
          type: number
          description: 对标准答案主线的覆盖度
          examples:
            - 72
        consistency_score:
          type: number
          description: 回答内部与标准答案的一致性
          examples:
            - 68
        completeness_score:
          type: number
          description: 对关键点和边界条件的完整度
          examples:
            - 61
      required: [coverage_score, consistency_score, completeness_score]
...